Optimize removeRedundantSlashes
It is often the case that the path string contains no redundant slashes.
When this is true, we still always create a new StringBuffer and result
string when we could just return the string itself (or the substring
that removes the trailing slash). On my machine, this reduced the
average runtime of this method from O(1us) to O(300ns). The extra
complexity is minimal so the optimization feels worth it.